# perf(core): use jemalloc as the memory allocator