--- a/packages/core/bin/run
+++ b/packages/core/bin/run
@@ -1,14 +1,22 @@
 #!/usr/bin/env node
 
-if (process.env.LD_PRELOAD !== 'libjemalloc.so') {
-    process.env.LD_PRELOAD = 'libjemalloc.so'
-    let exitStatus = 0
-    try {
-        require('child_process').execFileSync(process.argv0, process.argv.slice(1), { stdio: [0, 1, 2] })
-    } catch (error) {
-        exitStatus = error.status
+const child_process = require('child_process')
+
+const jemallocPath = child_process.spawnSync('whereis libjemalloc | xargs -n1 | grep libjemalloc.so', { shell: true }).stdout.toString().trim()
+
+if (jemallocPath) {
+    if (process.env.LD_PRELOAD !== jemallocPath) {
+        process.env.LD_PRELOAD = jemallocPath
+        let exitStatus = 0
+        try {
+            child_process.execFileSync(process.argv0, process.argv.slice(1), { stdio: [0, 1, 2] })
+        } catch (error) {
+            exitStatus = error.status
+        }
+        process.exit(exitStatus)
     }
-    process.exit(exitStatus)
+} else {
+    console.error('The jemalloc library was not found on your system. It is recommended to install it for better memory management. Falling back to the system default...')
 }
 
 require('@oclif/command').run()